requirements.txt
flask
flask-cors
waitress
sentence-transformers
langchain
transformers
faiss-cpu
torch
langchain-community